LPS-89415 Move to portal-apio
