--- modules/apps/headless-apio/rest-booster/rest-booster-apio-api/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay REST Booster Apio API
-Bundle-SymbolicName: com.liferay.rest.booster.apio.api
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.rest.booster.apio.context
