--- modules/apps/headless-apio/rest-booster/rest-booster-apio-api/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
-}
