--- modules/apps/headless-apio/rest-booster/rest-booster-apio-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay REST Booster Apio Implementation
-Bundle-SymbolicName: com.liferay.rest.booster.apio.impl
-Bundle-Version: 1.0.0
