--- modules/apps/headless-apio/rest-booster/rest-booster-apio-impl/build.gradle
@@ -1,19 +1,0 @@
-dependencies {
-	compileInclude group: "com.fasterxml.jackson.core", name: "jackson-annotations", version: "2.9.6"
-	compileInclude group: "com.fasterxml.jackson.core", name: "jackson-core", version: "2.9.6"
-	compileInclude group: "com.fasterxml.jackson.core", name: "jackson-databind", version: "2.9.6"
-
-	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.0"
-	compileOnly group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
-	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
-	compileOnly group: "org.apache.cxf", name: "cxf-core", version: "3.0.3"
-	compileOnly group: "org.apache.cxf", name: "cxf-rt-frontend-jaxrs", version: "3.0.0"
-	compileOnly group: "org.apache.cxf", name: "cxf-rt-rs-extension-providers", version: "3.0.3"
-	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	compileOnly group: "org.osgi", name: "org.osgi.service.jaxrs", version: "1.0.0"
-	compileOnly project(":apps:headless-apio:rest-booster:rest-booster-apio-api")
-}
