+++ modules/apps/portal-apio/portal-apio-api/src/main/resources/com/liferay/portal/apio/context/packageinfo
